Added exception test to wind and cell type memory allocation to catch allocation of too much memory and let us know. This also required making the nxdx, nydy, and nzdz variables longs since these values can get large.

+++ particleControl.h
@@ -183,7 +183,7 @@ class ParticleControl{
   int nz;
 
   float cell_dx,cell_dy,cell_dz;
-  int nxdx,nydy,nzdz;
+  long nxdx,nydy,nzdz;
 
   GLenum texType;
